Update README with ACA Viewer features and instructions
