+++ aca_viewer/__main__.py
@@ -1,0 +1,3 @@
+from . import run
+run()
+
